% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRunRes.R
\name{calcRunRes}
\alias{calcRunRes}
\title{Overload of calcRunRes() function to use inside shiny_encu()}
\usage{
calcRunRes(data_RFU)
}
\arguments{
\item{data_RFU}{\code{data.frame} uploaded data frame for calculating features on with no cycle column}
}
\value{
A data.frame with statisctics
}
\description{
Overload of calcRunRes() function to use inside shiny_encu()
}
