% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDataStatistics.R
\name{calculateDataStatistics}
\alias{calculateDataStatistics}
\title{Calculates statistics on given data and updates a progress bar}
\usage{
calculateDataStatistics(dataset)
}
\arguments{
\item{dataset}{\code{data.frame} uploaded data frame for calculating statistics on}
}
\value{
A character vector of length 1 with text to be displayed
}
\description{
Calculates statistics on given data and updates a progress bar
}
