% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_hash_maker.R
\name{dataset_hash_maker}
\alias{dataset_hash_maker}
\title{Calculates hash of a data set
Used to detect upload of a new datasets}
\usage{
dataset_hash_maker(dataset)
}
\arguments{
\item{dataset}{\code{data.frame} uploaded data frame for calculating hash on}
}
\value{
Hash of a dataset
}
\description{
Calculates hash of a data set
Used to detect upload of a new datasets
}
