% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_maker.R
\name{helper_maker}
\alias{helper_maker}
\title{Attaches a helper to a UI object.}
\usage{
helper_maker(fun, content)
}
\arguments{
\item{fun}{\code{function} UI object to have helper attached}

\item{content}{\code{character} A title of a helper.}
}
\value{
An UI object with helper attached
}
\description{
Attaches a helper to a UI object.
}
