% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectlySpecifiedColumnsAlert.R
\name{incorrectlySpecifiedColumnsAlert}
\alias{incorrectlySpecifiedColumnsAlert}
\title{Makes a shinyalert after fail of selecting columns to be displayed}
\usage{
incorrectlySpecifiedColumnsAlert(text, pasted)
}
\arguments{
\item{text}{\code{character} text to be displayed}

\item{pasted}{\code{character} text pasted}
}
\value{
Nothing
}
\description{
Makes a shinyalert after fail of selecting columns to be displayed
}
