% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncol_to_selectedColumns.R
\name{ncol_to_selectedColmns}
\alias{ncol_to_selectedColmns}
\title{Selects standard columns to display - first and last 5}
\usage{
ncol_to_selectedColmns(numCol)
}
\arguments{
\item{numCol}{\code{numeric} Number of solumns to select from}
}
\value{
Selected first and last 5 of columns
}
\description{
Selects standard columns to display - first and last 5
}
