% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurve.R
\name{plotCurve}
\alias{plotCurve}
\title{Generates a plot of a given column from a dataset}
\usage{
plotCurve(dataset, curvename)
}
\arguments{
\item{dataset}{\code{data.frame} uploaded data frame}

\item{curvename}{\code{character} name of one of columns of dataset; will be plotted}
}
\value{
A plotly graph
}
\description{
Generates a plot of a given column from a dataset
}
