% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_encu.R
\name{shiny_encu}
\alias{shiny_encu}
\title{Calculates features on given data and updates a progress bar}
\usage{
shiny_encu(data)
}
\arguments{
\item{data}{\code{data.frame} uploaded data frame for calculating features on}
}
\value{
A data.frame with features
}
\description{
Calculates features on given data and updates a progress bar
}
\note{
For updates of a progress bar to work properly, has to be placed
inside a withProgress() function.
}
