% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_extension.R
\name{extract_extension}
\alias{extract_extension}
\title{Extracts last extension of a file}
\usage{
extract_extension(file)
}
\arguments{
\item{file}{\code{character} file to extract extension from}
}
\value{
A last dot(.) and text after it
}
\description{
Extracts last extension of a file
}
